Modified the Launcher strategy
New Jobs are now submitted with the following strategy :
1. Create New JobID
2. Add into DB
3. Create directory
4. Return the JobID in the POST request
5. Continue on the server side by executing the query

Also, modified the Launcher shell script to update the Log and Status
file of the JobID based on the exit status of the Java Executor
program.

Also modified the Connection Establishment timeout to 10 seconds. This
will improve UX when Hive Server is down.